Move application code to tikki subdirectory
