+++ tikki/exceptions.py
